+++ agendamento/src/main/java/br/com/medcon/vo/TipoServico.java
@@ -2,20 +2,20 @@ package br.com.medcon.vo;
 
 
 public class TipoServico {
-    private Long id;
+    private int id;
     private String nome; 
     private int duracaoMinutos;
     private Especialidade especialidadeNecessaria; 
 
     public TipoServico() {}
-    public TipoServico(Long id, String nome, int duracaoMinutos, Especialidade especialidadeNecessaria) {
+    public TipoServico(int id, String nome, int duracaoMinutos, Especialidade especialidadeNecessaria) {
         this.id = id;
         this.nome = nome;
         this.duracaoMinutos = duracaoMinutos;
         this.especialidadeNecessaria = especialidadeNecessaria;
     }
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
     public int getDuracaoMinutos() { return duracaoMinutos; }
